requirements.txt
aiogram==2.25.1
emoji==2.6.0
fluent.runtime==0.4.0
